+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Alias !lmgtfy to !google
 
+### Changed
+- !version checks if this is the latest tag on GitHub too
+
 
 ## [0.3.0] - 2021-12-20
 ### Added